chore: refactor resolvers into separate file
