+++ src/app.js
@@ -3,37 +3,13 @@ import gql from 'graphql-tag'
 import { mergeSchemas } from 'graphql-tools'
 import { importSchema } from 'graphql-import'
 import { createGithubSchema, createGithubResolvers } from './github'
-import { ContributionAPI } from './contribution.provider'
+import { contributionResolvers } from './contribution.resolvers'
 
 const schema = importSchema(`./src/contribution.schema.graphql`)
-const api = new ContributionAPI()
 
 const localSchema = makeExecutableSchema({
   typeDefs: [schema],
-  resolvers: {
-    Query: {
-      async getContributions() {
-        return api.getAllContributions()
-      },
-      async getContribution(_, { input }) {
-        const { id } = input
-        return api.getContribution(id)
-      }
-    },
-    Mutation: {
-      async addContribution(_, { input }) {
-        return api.addContribution(input)
-      },
-      async updateContribution(_, { input }) {
-        const { id, newData } = input
-        return api.updateContribution(id, newData)
-      },
-      async deleteContribution(_, { input }) {
-        const { id } = input
-        return api.deleteContribution(id)
-      }
-    }
-  }
+  resolvers: contributionResolvers
 })
 
 const createSchema = async () => {
